LayoutTests/platform/win/imported/w3c/web-platform-tests/css/selectors/focus-visible-025-expected.txt

Initial
Target A
Target B

PASS ":focus-visible" should be a valid selector
FAIL :focus-visible matches after accesskey on DIV after previous mouse focus assert_equals: outline-style for DIV#targetA should be solid expected "solid" but got "none"
FAIL :focus-visible matches after accesskey on BUTTON after previous mouse focus assert_equals: outline-style for BUTTON#targetB should be solid expected "solid" but got "none"
